splits the two classes into their own files, and adjusts the main file to import the classes
